removed meta-data view and moved global constraints into licence policy item of rule editor
