+++ src/views/Start.vue
@@ -26,8 +26,7 @@ export default {
     return {
       modes: [
         { name: 'OER (Creative Commons)', path: { name: 'oer' } },
-        { name: 'Proprietär (RightsML)', path: { name: 'proprietary' } },
-        { name: 'Metadaten (zusätzlich zur Lizenz)', path: { name: 'meta' } },
+        { name: 'Proprietär (RightsML)', path: { name: 'proprietary' } }
       ],
     };
   },
